+++ stripe_terminal/android/src/main/kotlin/mek/stripeterminal/mappings/PaymentIntentMappings.kt
@@ -1,0 +1,10 @@
+package mek.stripeterminal.mappings
+
+import com.stripe.stripeterminal.external.models.AmountDetails
+import mek.stripeterminal.api.AmountDetailsApi
+
+fun AmountDetails.toApi(): AmountDetailsApi {
+    return AmountDetailsApi(
+        tip = tip?.toApi(),
+    )
+}