+++ stripe_terminal/ios/Classes/Mappings/CardMappings.swift
@@ -1,0 +1,15 @@
+import Foundation
+import StripeTerminal
+
+extension CardDetails {
+    func toApi() -> CardDetailsApi {
+        return CardDetailsApi(
+            brand: brand.toApi(),
+            country: country,
+            expMonth: expMonth,
+            expYear: expYear,
+            funding: funding.toApi(),
+            last4: last4
+        )
+    }
+}