+++ stripe_terminal/ios/Classes/Mappings/PaymentMethodMappings.swift
@@ -1,0 +1,15 @@
+import Foundation
+import StripeTerminal
+
+extension PaymentMethod {
+    func toApi() -> PaymentMethodApi {
+        return PaymentMethodApi(
+            id: stripeId,
+            card: card?.toApi(),
+            cardPresent: cardPresent?.toApi(),
+            interacPresent: interacPresent?.toApi(),
+            customerId: customer,
+            metadata: metadata
+        )
+    }
+}