+++ stripe_terminal/ios/Classes/Mappings/TipMappings.swift
@@ -1,0 +1,10 @@
+import Foundation
+import StripeTerminal
+
+extension SCPTip {
+    func toApi() -> TipApi {
+        return TipApi(
+            amount: amount?.intValue
+        )
+    }
+}