+++ stripe_terminal/lib/src/models/card.g.dart
@@ -6,6 +6,42 @@ part of 'card.dart';
 // DataClassGenerator
 // **************************************************************************
 
+mixin _$CardDetails {
+  CardDetails get _self => this as CardDetails;
+  @override
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+      other is CardDetails &&
+          runtimeType == other.runtimeType &&
+          _self.brand == other.brand &&
+          _self.country == other.country &&
+          _self.expMonth == other.expMonth &&
+          _self.expYear == other.expYear &&
+          _self.funding == other.funding &&
+          _self.last4 == other.last4;
+  @override
+  int get hashCode {
+    var hashCode = 0;
+    hashCode = $hashCombine(hashCode, _self.brand.hashCode);
+    hashCode = $hashCombine(hashCode, _self.country.hashCode);
+    hashCode = $hashCombine(hashCode, _self.expMonth.hashCode);
+    hashCode = $hashCombine(hashCode, _self.expYear.hashCode);
+    hashCode = $hashCombine(hashCode, _self.funding.hashCode);
+    hashCode = $hashCombine(hashCode, _self.last4.hashCode);
+    return $hashFinish(hashCode);
+  }
+
+  @override
+  String toString() => (ClassToString('CardDetails')
+        ..add('brand', _self.brand)
+        ..add('country', _self.country)
+        ..add('expMonth', _self.expMonth)
+        ..add('expYear', _self.expYear)
+        ..add('funding', _self.funding)
+        ..add('last4', _self.last4))
+      .toString();
+}
+
 mixin _$CardPresentDetails {
   CardPresentDetails get _self => this as CardPresentDetails;
   @override